Replacing the TodoListService by the same as in the active-directory-dotnet-native-aspnetcore (except the controller which does not need a user as this is a client credentials flow)
