--- TodoListService-Core/wwwroot/index.html
@@ -1,12 +1,0 @@
-﻿<!DOCTYPE html>
-<html>
-<head>
-    <meta charset="utf-8" />
-    <title></title>
-</head>
-<body>
-    This Web Api is running.
-    <p/>
-    Access /api/todolist from a client application
-</body>
-</html>
